@@ -1,26 +1,19 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import { useTheme } from 'hooks/use-theme';
+import { useEffect } from 'react';
+import Router from './router';
 
-function App() {
+export default function App() {
+  const { theme } = useTheme();
+
+  useEffect(() => {
+    document.body.classList.add(theme.theme, theme.color)
+
+    return () => {
+      document.body.classList.remove(theme.theme, theme.color);
+    }
+  }, [theme]);
+
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <Router />
   );
 }
-
-export default App;
